+++ inst/examples/getBoilerplateCode.r
@@ -4,7 +4,7 @@
 ## Important remarks //
 ##------------------------------------------------------------------------------
 ## This is a function that is currently only used inside a call to 'setReactive()'
-## or 'setThis_bare()'.
+## or 'setReactive_bare()'.
 ## Thus, in order to see what's going on here, we first need to mimick the 
 ## expected system state.